saveToQueue function finished, doctorId in QueueEntity changed referenced to DoctorFeatures(id), timetable modal updated

+++ src/main/java/com/example/medcenter/repoitory/QueueRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 
 @Repository
 public interface QueueRepository extends JpaRepository<QueueEntity, Long> {
-    List<QueueEntity> findQueueEntitiesByDateAndDoctorId(Date date , long doctorId);
+    List<QueueEntity> findQueueEntitiesByDateAndDoctorId(Date date , int doctorId);
 }
